requirements.txt
# Strands-Agents SDK Requirements
# Core Strands-Agents framework
strands-agents
strands-agents-tools

# Core dependencies
requests>=2.31.0
python-dateutil>=2.8.2

# Email functionality
secure-smtplib>=0.1.1

# Google Calendar API integration
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.88.0

# Optional model providers (choose based on your preference)
# OpenAI (if using OpenAI models)
openai>=1.0.0

# Anthropic (if using Claude models)
anthropic>=0.7.0

# AWS SDK (if using Amazon Bedrock - default for Strands-Agents)
boto3>=1.34.0
botocore>=1.34.0

# Additional utilities
python-dotenv>=1.0.0  # For environment variable management
colorama>=0.4.6       # For colored terminal output
rich>=13.0.0          # For enhanced terminal formatting

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0